Updated ComparableTestCase to extend HashCodeEqualsDefinedEqualityTestCase.
