+++ src/main/java/walkingkooka/test/HashCodeEqualsDefinedEqualityTestCase.java
