--- src/test/java/walkingkooka/net/IpPortEqualityTest.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2018 Miroslav Pokorny (github.com/mP1)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *
- */
-
-package walkingkooka.net;
-
-import org.junit.Test;
-import walkingkooka.test.HashCodeEqualsDefinedEqualityTestCase;
-
-public final class IpPortEqualityTest extends HashCodeEqualsDefinedEqualityTestCase<IpPort> {
-
-    private final static int PORT = 80;
-
-    @Test
-    public void testDifferentPort() {
-        this.checkNotEquals(IpPort.HTTPS);
-    }
-
-    @Test
-    public void testEqual() {
-        this.checkEqualsAndHashCode(IpPort.with(IpPortEqualityTest.PORT));
-    }
-
-    @Override
-    protected IpPort createObject() {
-        return IpPort.with(IpPortEqualityTest.PORT);
-    }
-}
