+++ src/test/java/walkingkooka/text/cursor/parser/spreadsheet/SpreadsheetCellReferenceComparableTest.java
@@ -18,6 +18,7 @@
 
 package walkingkooka.text.cursor.parser.spreadsheet;
 
+import org.junit.Ignore;
 import org.junit.Test;
 import walkingkooka.compare.LowerOrUpperTestCase;
 
@@ -51,9 +52,20 @@ public final class SpreadsheetCellReferenceComparableTest extends LowerOrUpperTe
         this.compareToAndCheckLess(this.cell(SpreadsheetReferenceKind.RELATIVE, COLUMN + 10, SpreadsheetReferenceKind.ABSOLUTE, ROW));
     }
 
+    @Test
+    @Ignore
+    public void testEqualsOnlyOverridesAbstractOrObject() {
+        throw new UnsupportedOperationException();
+    }
+
     @Override
     protected SpreadsheetCellReference createLowerOrUpper() {
         return this.cell(COLUMN, ROW);
+    }
+
+    @Override
+    protected boolean compareAndEqualsMatch() {
+        return false;
     }
 
     private SpreadsheetCellReference cell(final int column, final int row) {
@@ -65,10 +77,5 @@ public final class SpreadsheetCellReferenceComparableTest extends LowerOrUpperTe
                                           final SpreadsheetReferenceKind rowKind,
                                           final int row) {
         return columnKind.column(column).setRow(rowKind.row(row));
-    }
-
-    @Override
-    protected Class<SpreadsheetCellReference> type() {
-        return SpreadsheetCellReference.class;
     }
 }
